+++ lib/WeekView.dart
@@ -1,11 +1,11 @@
-
 import 'package:flutter/material.dart';
 import 'main.dart';
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'Task.dart';
 import 'Occurrence.dart';
 import 'User.dart';
-
+import 'main.dart';
+import 'MyHomepage.dart';
 
 class WeekView extends StatelessWidget {
   @override
@@ -36,7 +36,7 @@ class _MyWeekViewState extends State<MyWeekView> {
 
   void createAllTaskData() {
     User myUser = MyHomePage.theUser;
-    
+
     for (int i = 0; i < 7; i++) {
       List<TaskData> listOfDayTasks = [];
       Map<String, int> listTasks = new Map();
@@ -65,9 +65,7 @@ class _MyWeekViewState extends State<MyWeekView> {
       String name = "";
 
       for (int i = 0; i < 24; i++) {
-        
         if (startTimes.contains(i)) {
-
           for (Task task in myUser.tasks) {
             for (Occurrence occurs in task.occurrence) {
               if (occurs.time[0].hour == i) {
@@ -77,16 +75,16 @@ class _MyWeekViewState extends State<MyWeekView> {
           }
           if (i != end) {
             if (end < i) {
-              listOfDayTasks.add(new TaskData('free time', i-end));
+              listOfDayTasks.add(new TaskData('free time', i - end));
             }
           }
           end = listEndTimes[name];
           start = i;
-          listOfDayTasks.add(new TaskData(name, end-start));
+          listOfDayTasks.add(new TaskData(name, end - start));
         }
         if (i == 23) {
           if (i != end) {
-            listOfDayTasks.add(new TaskData('free time', i-end));
+            listOfDayTasks.add(new TaskData('free time', i - end));
           }
         }
       }
@@ -97,24 +95,13 @@ class _MyWeekViewState extends State<MyWeekView> {
       listOfAllDays.add(listOfDayTasks);
     }
 
-
-
-
-
-
-
-
-
-
-
-
-
     setState() {
       build(context);
     }
   }
+
   List<List<charts.Series<TaskData, int>>> listOfAllData = [];
-  
+
   void data() {
     for (int i = 0; i < 7; i++) {
       listOfAllData.add(mondayData(i));
@@ -134,6 +121,7 @@ class _MyWeekViewState extends State<MyWeekView> {
       )
     ];
   }
+
 // child: DonutPieChart(mondayData()),
   @override
   Widget build(BuildContext context) {
@@ -149,187 +137,199 @@ class _MyWeekViewState extends State<MyWeekView> {
           child: Column(
             children: <Widget>[
               Container(
-                child: Center(
+                  child: Center(
                 child: Container(
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: <Widget>[
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Monday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[0]),
-                            ),
-                          )
-                        ],
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Monday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[0]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[0], 0)),
+                          );
+                        },
                       ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[0], 0)),
-                      );
-                    },
-                    ),
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Tuesday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[1]),
-                            ),
-                          )
-                        ],
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Tuesday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[1]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[1], 1)),
+                          );
+                        },
                       ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[1], 1)),
-                      );
-                    },
-                    ),
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Wednesday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[2]),
-                            ),
-                          )
-                        ],
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Wednesday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[2]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[2], 2)),
+                          );
+                        },
                       ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[2], 2)),
-                      );
-                    },
-                    ),
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Thursday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[3]),
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[3], 3)),
-                      );
-                    },
-                    )
-                  ],
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Thursday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[3]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[3], 3)),
+                          );
+                        },
+                      )
+                    ],
+                  ),
                 ),
-              ),
-                )
-              ),
+              )),
               Container(
-                child: Center(
+                  child: Center(
                 child: Container(
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: <Widget>[
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Friday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[4]),
-                            ),
-                          )
-                        ],
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Friday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[4]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[4], 4)),
+                          );
+                        },
                       ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[4], 4)),
-                      );
-                    },
-                    ),
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Saturday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[5]),
-                            ),
-                          )
-                        ],
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Saturday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[5]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[5], 5)),
+                          );
+                        },
                       ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[5], 5)),
-                      );
-                    },
-                    ),
-                    GestureDetector(
-                      child: Container(
-                      child: Column(
-                        children: <Widget>[
-                          Text('Sunday'),
-                        Container(
-                            height: 80,
-                            width: 80,
-                              child: Center(
-                              child: DonutPieChart(listOfAllData[6]),
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                    onTap: () {
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(builder: (context) => DayView(listOfAllData[6], 6)),
-                      ).then((value) {
-                        setState(() {
-                         createAllTaskData();
-                         build(context);
-                         print("HELLO" + value);                
-                      });
-                      });
-                    },
-                    )
-                  ],
+                      GestureDetector(
+                        child: Container(
+                          child: Column(
+                            children: <Widget>[
+                              Text('Sunday'),
+                              Container(
+                                height: 80,
+                                width: 80,
+                                child: Center(
+                                  child: DonutPieChart(listOfAllData[6]),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                        onTap: () {
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) =>
+                                    DayView(listOfAllData[6], 6)),
+                          ).then((value) {
+                            setState(() {
+                              createAllTaskData();
+                              build(context);
+                              print("HELLO" + value);
+                            });
+                          });
+                        },
+                      )
+                    ],
+                  ),
                 ),
-              ),
-                )
-              )
+              ))
             ],
           ),
         ),
@@ -338,17 +338,17 @@ class _MyWeekViewState extends State<MyWeekView> {
         heroTag: "Hellooo",
         child: Icon(Icons.add),
         onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => MyApp()),
-            );
-              setState() {
-                build(context);
-              }
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => MyHomePage(title: 'Timex')),
+          );
+          setState() {
+            build(context);
+          }
         },
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
-   );
+    );
   }
 }
 
@@ -366,7 +366,6 @@ class DonutPieChart extends StatelessWidget {
       animate: false,
     );
   }
-
 
   @override
   Widget build(BuildContext context) {
@@ -405,7 +404,6 @@ class TaskData {
   TaskData(this.name, this.time);
 }
 
-
 class DayView extends StatefulWidget {
   List<charts.Series<TaskData, int>> listOfTasks;
   int day;
@@ -418,21 +416,21 @@ class DayView extends StatefulWidget {
 }
 
 class _DayViewState extends State<DayView> {
-   List<charts.Series<TaskData, int>> listOfTasks;
-   Map<int, String> days = new Map();
-   int day;
-   
+  List<charts.Series<TaskData, int>> listOfTasks;
+  Map<int, String> days = new Map();
+  int day;
 
   _DayViewState(List<charts.Series<TaskData, int>> list, int day) {
     this.listOfTasks = list;
     this.day = day;
   }
 
-   setIt() {
+  setIt() {
     setState() {
       build(context);
     }
   }
+
   BuildContext theContext;
   @override
   Widget build(BuildContext context) {
@@ -446,12 +444,13 @@ class _DayViewState extends State<DayView> {
     days[6] = 'Sunday';
 
     return Scaffold(
-      appBar: AppBar(
-        title: Text(days[day],
-      ),
-      ),
-      body: Center(
-        child: Column(
+        appBar: AppBar(
+          title: Text(
+            days[day],
+          ),
+        ),
+        body: Center(
+            child: Column(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
             Container(
@@ -462,16 +461,15 @@ class _DayViewState extends State<DayView> {
             Container(
               constraints: BoxConstraints.expand(height: 260),
               child: ListView(
-                  children: makeTaskList(),
+                children: makeTaskList(),
               ),
             )
           ],
-        )
-      )
-    );
+        )));
   }
 
-  TextStyle labelStyle = new TextStyle(fontFamily: 'Avenir', fontSize: 20, fontWeight: FontWeight.bold);
+  TextStyle labelStyle = new TextStyle(
+      fontFamily: 'Avenir', fontSize: 20, fontWeight: FontWeight.bold);
 
   List<Widget> makeTaskList() {
     List<String> taskNames = [];
@@ -492,39 +490,50 @@ class _DayViewState extends State<DayView> {
     List<Widget> rows = [];
     for (int i = 0; i < taskNames.length; i++) {
       Container newContainer = new Container(
-        
-        height: 100,
-        width: 300,
-        child: GestureDetector(
-          
-          child: Card(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
+          height: 100,
+          width: 300,
+          child: GestureDetector(
+            child: Card(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Text(taskNames[i], style: labelStyle,),
-                  Text(makeTheTimeString(timeOfDay[taskNames[i]], hoursList[taskNames[i]]), style: labelStyle,),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    children: <Widget>[
+                      Text(
+                        taskNames[i],
+                        style: labelStyle,
+                      ),
+                      Text(
+                        makeTheTimeString(
+                            timeOfDay[taskNames[i]], hoursList[taskNames[i]]),
+                        style: labelStyle,
+                      ),
+                    ],
+                  ),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    children: <Widget>[
+                      Text(
+                        "Total Time: ",
+                        style: labelStyle,
+                      ),
+                      Text(
+                        hoursList[taskNames[i]].toString() + "hrs",
+                        style: labelStyle,
+                      ),
+                    ],
+                  )
                 ],
               ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: <Widget>[
-                  Text("Total Time: ", style: labelStyle,),
-                  Text(hoursList[taskNames[i]].toString() + "hrs", style: labelStyle,),
-                ],
-              )
-            ],
-          ),
-          elevation: 5,
-        ),
-        onTap: () {
-          print("Pressed");
-          _showDialog(taskNames[i], hoursList[taskNames[i]], timeOfDay[taskNames[i]]);
-        },
-        )
-      );
+              elevation: 5,
+            ),
+            onTap: () {
+              print("Pressed");
+              _showDialog(taskNames[i], hoursList[taskNames[i]],
+                  timeOfDay[taskNames[i]]);
+            },
+          ));
       rows.add(newContainer);
     }
     return rows;
@@ -538,25 +547,24 @@ class _DayViewState extends State<DayView> {
         // return object of type Dialog
         return AlertDialog(
           title: new Text(taskName),
-          content: new Text('Time: ' + makeTheTimeString(startTimeOfDay, duration)),
+          content:
+              new Text('Time: ' + makeTheTimeString(startTimeOfDay, duration)),
           actions: <Widget>[
             // usually buttons at the bottom of the dialog
             new FlatButton(
-
               child: new Text("Close"),
               onPressed: () {
                 Navigator.of(context).pop();
               },
             ),
             new FlatButton(
-              key: new Key('one'),
-              child: new Text("Delete"),
-              onPressed: () {
-                deleteATask(taskName);
-                Navigator.of(context).pop();
-                Navigator.of(theContext).pop();
-              }
-            ),
+                key: new Key('one'),
+                child: new Text("Delete"),
+                onPressed: () {
+                  deleteATask(taskName);
+                  Navigator.of(context).pop();
+                  Navigator.of(theContext).pop();
+                }),
           ],
         );
       },
@@ -575,23 +583,21 @@ class _DayViewState extends State<DayView> {
 
   String makeTheTimeString(DateTime date, int duration) {
     if (date.minute > 10) {
-      return date.hour.toString() + ":" + date.minute.toString() + "-" + (date.hour + duration).toString() + ":" + date.minute.toString();
+      return date.hour.toString() +
+          ":" +
+          date.minute.toString() +
+          "-" +
+          (date.hour + duration).toString() +
+          ":" +
+          date.minute.toString();
     } else {
-      return date.hour.toString() + ":0" + date.minute.toString() + "-" + (date.hour + duration).toString() + ":0" + date.minute.toString();
+      return date.hour.toString() +
+          ":0" +
+          date.minute.toString() +
+          "-" +
+          (date.hour + duration).toString() +
+          ":0" +
+          date.minute.toString();
     }
   }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
